fix issue where any item update would mark an item sold, and modularize item services
