--- app/services/item/expirer.rb
@@ -1,9 +1,0 @@
-class Item::Expirer
-
-  attr_reader :item
-
-  def expire!(items)
-    items.map(&:mark_expired)
-  end
-
-end
