+++ app/services/items/expirer.rb
@@ -1,0 +1,11 @@
+module Items
+  class Expirer
+
+    attr_reader :item
+
+    def expire!(items)
+      items.map(&:mark_expired)
+    end
+
+  end
+end
