+++ spec/services/items/expirer_spec.rb
@@ -1,0 +1,20 @@
+module Items
+  describe Expirer do
+
+    let!(:items) { create_list(:item, 2, :active, :consigned, listed_at: 93.days.ago) }
+    let(:expirer) { Items::Expirer.new }
+
+    it "can be instantiated" do
+      expect(expirer).to be_an_instance_of(Items::Expirer)
+    end
+
+    it "expires items" do
+      expect {
+        expirer.expire!(items)
+      }.to change {
+        Item.expired.count
+      }.by(2)
+    end
+
+  end
+end
